+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Fluent interface coverage has been increased to now include remove methods and parsing interfaces.
+- ShlexParser.add_option and add_argument methods which mirror the behaviour of `with_option` and `with_argument`.
+
+### Removed
+- add_converter and remove_converter from Parameter.
 
 ## [2.1.0a1] - 2021-10-02
 ### Added